+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- `OpenModelica.Scripting.CheckSettingsResult`
+    - Rename `SENDDATALIBS` => `RTLIBS`
+
 ## [0.3.0a4] - 2023-11-22
 
 ### Added
@@ -84,28 +89,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Update modelica parser library to `modelicalang>=0.1`
 
 ## [0.1.1] - 2021-12-13
+
 ### Added
+
 - `pyproject.toml` to unify package settings.
 - `omc4py.v_1_18` support of _OpenModelica 1.18.x_
 
 ### Removed
+
 - `setup.py`, `setup.cfg`, `requirements.txt` (unified to `pyproject.toml`)
 
 ## [0.1.0] - 2021-04-22
+
 ### Added
+
 - `omc4py.v_1_17` Support of _OpenModelica 1.17.x_
 - `omc4py.__version__` version string s.t. `"0.1.0"`
 - CHANGELOG.md file for version management
 - tests/ directory for testcases using pytest
+
 ### Removed
+
 - `omc4py.v_1_14.OMCSession.oms_faultInjection` not available due to a bug in omc 1.14.1
 
 ## [0.0.1] - 2020-11-29
+
 ### Fixed
+
 - setup.py fixed so that omc4py can be installed correctly
 
 ## 0.0.0 - 2020-11-29 [YANKED]
+
 ### Added
+
 - COPYING file for OSMC-PL license
 - omc4py/ directory (python package) for main package
 - bootstrap/ directory (python package) for bootstrap of omc4py
@@ -113,7 +129,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - MANIFEST.in file for setup.py
 - README.md file for tutorial
 - TODO.md file for action item management
+
 ### Security
+
 - __This version cannot be installed with setup.py__ because the packages required to run are not listed correctly.
 
 [Unreleased]: https://github.com/ijknabla/OpenModelicaCompilerForPython/compare/v0.3.0a4...HEAD